add middleware

+++ routes/web.php
@@ -41,27 +41,33 @@ Route::get('/contact', function () {
 //CRUD 
 // 3Routing :create/edit/list
 
-//C create controller 的儲存操作
-Route::post('/posts','PostController@store');
-//管理者頁面
-Route::get('/posts/admin','PostController@admin');
-//1 add read routing 
-Route::get('/posts/show/{post}','PostController@show');
 
+//管理者頁面
+Route::get('/posts/admin','PostController@admin')->middleware('auth');
 //create 表單
-Route::get('/posts/create','PostController@create');
+Route::get('/posts/create','PostController@create')->middleware('auth');
+//edit form
+Route::get('/posts/{post}/edit','PostController@edit')->middleware('auth');
+
+
+
+
+//C create controller 的儲存操作
+Route::post('/posts','PostController@store')->middleware('auth');
+//1 add read routing 
+Route::get('/posts/show/{post}','PostController@show')->middleware('auth');
+
+Route::put('/posts/{post}','PostController@update')->middleware('auth');
+//D
+Route::delete('/posts/{post}','PostController@destroy')->middleware('auth');
+
+
+//for everyone 
+//list
+Route::get('/posts','PostController@index');
 //R read controller 的儲存操作 。{post}指的是post 的model 最後會疊代成數字@後面為method 的名子 只要兩邊對應就可以
 Route::get('/posts/{post}','PostController@show');
 //Update 跟資料庫互動
-Route::put('/posts/{post}','PostController@update');
-//D
-Route::delete('/posts/{post}','PostController@destroy');
-
-
-//edit form
-Route::get('/posts/{post}/edit','PostController@edit');
-//list
-Route::get('/posts','PostController@index');
 
 
 Auth::routes();
